+++ app/api/__init__.py
@@ -1,5 +1,6 @@
 from fastapi import APIRouter
 
+from app.api.v1.app_settings import router as app_settings_router
 from app.api.v1.auth import router as auth_router
 from app.api.v1.client_application import router as client_application_router
 from app.api.v1.introspect import router as introspect_router
@@ -10,10 +11,11 @@ from app.api.v1.user import router as user_router
 
 api_router = APIRouter()
 
-api_router.include_router(oidc_router, prefix="/oidc", tags=["oidc"])
-api_router.include_router(auth_router, tags=["auth"])
-api_router.include_router(token_router, tags=["token"])
+api_router.include_router(oidc_router, prefix="/oidc", tags=["OIDC"])
+api_router.include_router(auth_router, tags=["Auth"])
+api_router.include_router(token_router, tags=["Token"])
 api_router.include_router(client_application_router, tags=["Client Applications"])
-api_router.include_router(introspect_router, tags=["introspect"])
-api_router.include_router(revoke_router, tags=["revoke"])
-api_router.include_router(user_router, tags=["user"])
+api_router.include_router(introspect_router, tags=["Introspect"])
+api_router.include_router(revoke_router, tags=["Revoke"])
+api_router.include_router(user_router, tags=["User"])
+api_router.include_router(app_settings_router, tags=["App Settings"])