+++ app/models/__init__.py
@@ -2,3 +2,4 @@ from .refresh_token import RefreshToken
 from .access_token import AccessToken
 from .user import User
 from .client_application import ClientApplication
+from .app_settings import AppSetting